Make loadLocalShard public, add a ton of indirection to sidestep circular import of db package
