+++ adapters/repos/db/index.go
@@ -1979,13 +1979,8 @@ func (i *Index) initLocalShard(ctx context.Context, shardName string) error {
 	return i.initLocalShardWithForcedLoading(ctx, i.getClass(), shardName, false)
 }
 
-func (i *Index) loadLocalShard(ctx context.Context, shardName string) error {
+func (i *Index) LoadLocalShard(ctx context.Context, shardName string) error {
 	return i.initLocalShardWithForcedLoading(ctx, i.getClass(), shardName, true)
-}
-
-// TODO temp for testing
-func (i *Index) LoadShard(ctx context.Context, shardName string) error {
-	return i.loadLocalShard(ctx, shardName)
 }
 
 func (i *Index) initLocalShardWithForcedLoading(ctx context.Context, class *models.Class, shardName string, mustLoad bool) error {
