+++ cluster/replication/copier/types/types.go
@@ -1,0 +1,33 @@
+package types
+
+import (
+	"context"
+	"io"
+
+	"github.com/weaviate/weaviate/adapters/repos/db"
+	"github.com/weaviate/weaviate/entities/schema"
+)
+
+// IndexGetter is a type that can get an index, this is used to avoid a circular
+// dependency between the copier and the db package.
+type IndexGetter interface {
+	GetIndex(name schema.ClassName) *db.Index
+}
+
+// ShardLoader is a type that can load a shard from disk files, this is used to avoid a circular
+// dependency between the copier and the db package.
+type ShardLoader interface {
+	// LoadLocalShard See adapters/repos/db.Index.LoadLocalShard
+	LoadLocalShard(ctx context.Context, name string) error
+}
+
+// RemoteIndex is a type that can interact with a remote index, this is used to avoid a circular
+// dependency between the copier and the db package.
+type RemoteIndex interface {
+	// PauseAndListFiles See adapters/clients.RemoteIndex.PauseAndListFiles
+	PauseAndListFiles(ctx context.Context,
+		hostName, indexName, shardName string) ([]string, error)
+	// GetFile See adapters/clients.RemoteIndex.GetFile
+	GetFile(ctx context.Context,
+		hostName, indexName, shardName, fileName string) (io.ReadCloser, error)
+}
